turn video grid into a component (wip)

+++ frontend/src/routes/library/videos.tsx
@@ -1,12 +1,10 @@
 import {useStateMachine} from "little-state-machine"
 import {
   HiChevronRight,
-  HiFolder,
   HiOutlineArrowPath,
   HiPlus,
   HiTag,
   HiTrash,
-  HiXMark,
 } from "react-icons/hi2"
 import {useState} from "react"
 import {updateForm} from "../actions"
@@ -15,7 +13,6 @@ import {
   Outlet,
   useLoaderData,
   useNavigate,
-  useNavigation,
   useRevalidator,
   useSearchParams,
 } from "react-router-dom"
@@ -24,73 +21,24 @@ import {
   cleanupVideos,
   deleteVideo,
   mergeStashVideo,
-  updateVideo,
 } from "../../api"
-import Pagination from "../../components/Pagination"
 import {FormStage} from "../../types/form-state"
-import VideoCard from "../../components/VideoCard"
-import {useConfig} from "../../hooks/useConfig"
-import useDebouncedSetQuery from "../../hooks/useDebouncedQuery"
+
 import JumpToTop from "../../components/JumpToTop"
+import VideoGrid from "@/components/VideoGrid"
 
 export default function ListVideos() {
   const {actions} = useStateMachine({updateForm})
-  const initialVideos = useLoaderData() as ListVideoDtoPage
-  const videos = initialVideos.content
   const navigate = useNavigate()
-  const navigation = useNavigation()
-  const config = useConfig()
-  const isLoading = navigation.state === "loading"
-  const [params] = useSearchParams()
-  const [source, setSource] = useState(params.get("source") || undefined)
-  const [filter, setFilter] = useState(params.get("query") ?? "")
-  const noVideos = videos.length === 0 && !filter && !isLoading && !source
+  const page = useLoaderData() as ListVideoDtoPage
   const revalidator = useRevalidator()
-  const [sort, setSort] = useState(params.get("sort") ?? "markers")
-  const {setQueryDebounced, addOrReplaceParam} = useDebouncedSetQuery()
   const [syncingVideo, setSyncingVideo] = useState<string>()
-  const [hasMarkers, setHasMarkers] = useState<string | undefined>(
-    params.get("hasMarkers") ?? "",
-  )
+  const videos = page.content
   const [query] = useSearchParams()
-  const anyFilterSet = filter || source || hasMarkers
-
-  const noVideosForFilter = videos.length === 0 && anyFilterSet && !isLoading
 
   const onOpenModal = (videoId: string) => {
     const queryString = query.toString()
     navigate(`/library/${videoId}/markers?${queryString}`)
-  }
-
-  const onFilterChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setFilter(e.target.value)
-    setQueryDebounced(e.target.value.trim())
-  }
-
-  const onChangeSort = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    setSort(e.target.value)
-    addOrReplaceParam("sort", e.target.value)
-  }
-
-  const onChangeHasMarkers = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    const value = e.target.value
-    setHasMarkers(value)
-    if (value === "") {
-      addOrReplaceParam("hasMarkers", undefined)
-    } else {
-      addOrReplaceParam("hasMarkers", value)
-    }
-  }
-
-  const onChangeSource = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    const value = e.target.value
-    if (value === "All") {
-      setSource(undefined)
-      addOrReplaceParam("source", undefined)
-    } else {
-      setSource(value)
-      addOrReplaceParam("source", value)
-    }
   }
 
   const onNextStage = () => {
@@ -107,7 +55,9 @@ export default function ListVideos() {
   }
 
   const onRemoveVideo = async (id: string) => {
-    if (confirm("Are you sure you want to remove this video?")) {
+    if (
+      confirm("Are you sure you want to remove this video and all its markers?")
+    ) {
       await deleteVideo(id)
       revalidator.revalidate()
     }
@@ -125,10 +75,6 @@ export default function ListVideos() {
     }
   }
 
-  const onEditTitle = async (id: string, title: string) => {
-    await updateVideo(id, {title})
-  }
-
   const onSyncVideo = async (id: string) => {
     setSyncingVideo(id)
     await mergeStashVideo(id)
@@ -136,8 +82,8 @@ export default function ListVideos() {
     setSyncingVideo(undefined)
   }
 
-  const startRange = initialVideos.pageNumber * initialVideos.pageSize + 1
-  const endRange = startRange + initialVideos.content.length - 1
+  const startRange = page.pageNumber * page.pageSize + 1
+  const endRange = startRange + page.content.length - 1
 
   return (
     <>
@@ -162,7 +108,7 @@ export default function ListVideos() {
             <strong>
               {startRange}-{endRange}
             </strong>{" "}
-            of <strong>{initialVideos.totalItems}</strong>.
+            of <strong>{page.totalItems}</strong>.
           </p>
         </div>
         {videos.length > 0 && (
@@ -175,141 +121,40 @@ export default function ListVideos() {
           </button>
         )}
       </div>
-      {!noVideos && (
-        <div className="w-full grid grid-cols-2">
-          <input
-            type="text"
-            placeholder="Filter..."
-            className="input input-primary w-full lg:w-96"
-            value={filter}
-            onChange={onFilterChange}
-          />
 
-          <div className="flex gap-2 place-self-end">
+      <VideoGrid
+        editableTitles
+        onVideoClick={onOpenModal}
+        actionChildren={(video) => (
+          <>
+            <div className="form-control"></div>
             <div className="flex gap-1">
-              <label className="label">
-                <span className="label-text">Video source</span>
-              </label>
-              <select
-                value={source}
-                onChange={onChangeSource}
-                className="select select-sm select-bordered"
+              {video.video.source === "Stash" && (
+                <button
+                  onClick={() => onSyncVideo(video.video.id)}
+                  className="btn btn-sm btn-secondary"
+                  disabled={syncingVideo === video.video.id}
+                >
+                  <HiOutlineArrowPath />
+                  Sync
+                </button>
+              )}
+              <button
+                onClick={() => onRemoveVideo(video.video.id)}
+                className="btn btn-error btn-sm btn-square"
               >
-                <option disabled value="none">
-                  Filter video source
-                </option>
-                <option value="All">All</option>
-                <option value="Folder">Local folder</option>
-                <option value="Stash">Stash</option>
-                <option value="Download">Downloaded</option>
-              </select>
-            </div>
-
-            <div className="flex gap-1">
-              <label className="label">
-                <span className="label-text">Sort by</span>
-              </label>
-              <select
-                value={sort}
-                onChange={onChangeSort}
-                className="select select-sm select-bordered"
+                <HiTrash />
+              </button>
+              <button
+                onClick={() => onOpenModal(video.video.id)}
+                className="btn btn-sm btn-primary"
               >
-                <option disabled value="none">
-                  Sort by...
-                </option>
-                <option value="markers">Number of markers</option>
-                <option value="title">Video title</option>
-                <option value="created">Created on</option>
-                <option value="duration">Duration</option>
-              </select>
+                <HiTag className="w-4 h-4 mr-2" />
+                Markers
+              </button>
             </div>
-
-            <div className="flex gap-1">
-              <label className="label">
-                <span className="label-text">Has markers</span>
-              </label>
-              <select
-                value={hasMarkers}
-                onChange={onChangeHasMarkers}
-                className="select select-sm select-bordered"
-              >
-                <option disabled value="none">
-                  Has markers
-                </option>
-                <option value="">All</option>
-                <option value="true">Yes</option>
-                <option value="false">No</option>
-              </select>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {noVideos && (
-        <div className="flex flex-col items-center justify-center mt-8">
-          <HiFolder className="text-8xl" />
-          <h1 className="text-xl">No videos found</h1>
-          <p>Add some videos with the button above!</p>
-        </div>
-      )}
-
-      {noVideosForFilter && (
-        <div className="flex flex-col items-center justify-center mt-8">
-          <HiXMark className="text-8xl" />
-          <h1 className="text-xl">No videos found</h1>
-          <p>Try changing the filter.</p>
-        </div>
-      )}
-
-      <section className="grid grid-cols-1 lg:grid-cols-3 gap-4 w-full my-4">
-        {videos.map((video) => (
-          <VideoCard
-            key={video.video.id}
-            onImageClick={onOpenModal}
-            video={video}
-            stashConfig={config}
-            onEditTitle={
-              video.video.source === "Stash"
-                ? undefined
-                : (title) => onEditTitle(video.video.id, title)
-            }
-            actionChildren={
-              <>
-                <div className="form-control"></div>
-                <div className="flex gap-1">
-                  {video.video.source === "Stash" && (
-                    <button
-                      onClick={() => onSyncVideo(video.video.id)}
-                      className="btn btn-sm btn-secondary"
-                      disabled={syncingVideo === video.video.id}
-                    >
-                      <HiOutlineArrowPath />
-                      Sync
-                    </button>
-                  )}
-                  <button
-                    onClick={() => onRemoveVideo(video.video.id)}
-                    className="btn btn-error btn-sm btn-square"
-                  >
-                    <HiTrash />
-                  </button>
-                  <button
-                    onClick={() => onOpenModal(video.video.id)}
-                    className="btn btn-sm btn-primary"
-                  >
-                    <HiTag className="w-4 h-4 mr-2" />
-                    Markers
-                  </button>
-                </div>
-              </>
-            }
-          />
-        ))}
-      </section>
-
-      <Pagination
-        totalPages={initialVideos.totalPages}
-        currentPage={initialVideos.pageNumber}
+          </>
+        )}
       />
     </>
   )
